+++ src/index.ts
@@ -1,7 +1,6 @@
-export type CallbackFieldValueType = string;
-export type CallbackFieldKey = string;
-export type FunctionCallbackArg = unknown;
-export type OnSuccessPayloadResult = any;
+export type CallbackEnding = string;
+export type ExecutionFunction = Function;
+export type NonFunctionFieldValue = unknown;
 export enum ProcessingType {
   synchronous = "synchronous",
   promise = "promise async",
@@ -12,12 +11,12 @@ export enum ProcessingResult {
   failed = "failed",
 }
 export interface CallbackPayload {
-  fieldValueType: CallbackFieldValueType;
-  fieldKey: CallbackFieldKey;
+  fieldValueType: string;
+  fieldKey: string;
 }
 export interface FunctionCallbackPayload {
   processingStrategy: ProcessingType;
-  functionArgs: FunctionCallbackArg[];
+  functionArgs: unknown[];
 }
 export interface OnErrorAdditionalPayload {
   processingResult: ProcessingResult.failed;
@@ -25,64 +24,71 @@ export interface OnErrorAdditionalPayload {
 }
 export interface OnSuccessAdditionalPayload {
   processingResult: ProcessingResult.succeed;
-  functionResult: OnSuccessPayloadResult;
+  functionResult: unknown;
 }
 export type OnErrorPayload = CallbackPayload & FunctionCallbackPayload & OnErrorAdditionalPayload;
 export type OnSuccessPayload = CallbackPayload & FunctionCallbackPayload & OnSuccessAdditionalPayload;
 export type OnNonFunctionPayload = CallbackPayload;
-export type OnError = (payload: OnErrorPayload) => Error | void;
-export type OnSuccess = (payload: OnSuccessPayload) => void;
-export type OnNonFunction = (payload: OnNonFunctionPayload) => void;
+export type OnErrorResult = OnErrorAdditionalPayload["functionError"] | void;
+export type OnSuccessResult = void;
+export type OnNonFunctionResult = void;
+export type OnError = (payload: OnErrorPayload) => OnErrorResult;
+export type OnSuccess = (payload: OnSuccessPayload) => OnSuccessResult;
+export type OnNonFunction = (payload: OnNonFunctionPayload) => OnNonFunctionResult;
 
-/**
- * @description
- *`callbackEnding` this field is used when an object (on which this proxy handler is applied) function use callback style to return value and have callback to promise transformation function. This field define a name of a callback to promise transformation function which should be executed after primary function execution.
- * @example
- * ```
- * import { StepFunctions } from "aws-sdk";
- * import { proxyHandlerGenericExecution } from "proxy-handler-generic-execution";
- *
- * const callbackEnding = "promise";
- * const stepFunctions = new StepFunctions();
- * const wrappedStepFunctions = new Proxy(
- *   stepFunctions,
- *   proxyHandlerGenericExecution({
- *     callbackEnding,
- *     onSuccess: () => {},
- *     onNonFunction: () => {},
- *     onError: () => {},
- *   }),
- * );
- *
- * (async () => {
- *   await wrappedStepFunctions.startExecution({ stateMachineArn: "ARN" })[callbackEnding]();
- * })();
- * ```
- */
-
-export type ProxyHandlerGenericExecutionOptions<CallbackEnding = string | undefined> = {
+export type ProxyHandlerGenericExecutionOptions = {
   onError: OnError;
   onSuccess: OnSuccess;
   onNonFunction: OnNonFunction;
-} & (CallbackEnding extends undefined ? { callbackEnding: string } : any);
+  /**
+   * @description
+   *`callbackEnding` this field is used when an object (on which this proxy handler is applied) function use callback style to return value and have callback to promise transformation function. This field define a name of a callback to promise transformation function which should be executed after primary function execution.
+   * @example
+   * ```
+   * import { StepFunctions } from "aws-sdk";
+   * import { proxyHandlerGenericExecution } from "proxy-handler-generic-execution";
+   *
+   * const callbackEnding = "promise";
+   * const stepFunctions = new StepFunctions();
+   * const wrappedStepFunctions = new Proxy(
+   *   stepFunctions,
+   *   proxyHandlerGenericExecution({
+   *     callbackEnding,
+   *     onSuccess: () => {},
+   *     onNonFunction: () => {},
+   *     onError: () => {},
+   *   }),
+   * );
+   *
+   * (async () => {
+   *   await wrappedStepFunctions.startExecution({ stateMachineArn: "ARN" })[callbackEnding]();
+   * })();
+   * ```
+   */
+  callbackEnding?: CallbackEnding;
+};
 
 export type ProxyHandlerGenericExecution = (options: ProxyHandlerGenericExecutionOptions) => ProxyHandler<any>;
 
 export const proxyHandlerGenericExecution: ProxyHandlerGenericExecution = (options) => ({
   get: (target, key) => {
-    const fieldValueType = typeof target[key];
+    const fieldValueType: CallbackPayload["fieldValueType"] = typeof target[key];
     if (fieldValueType !== "function") {
       options.onNonFunction({ fieldKey: String(key), fieldValueType });
-      return target[key];
+      return target[key] as NonFunctionFieldValue;
     }
+
     return (...functionArgs: FunctionCallbackPayload["functionArgs"]) => {
       try {
-        const syncResult = target[key](...functionArgs);
+        const syncResult: unknown | Promise<unknown> | (unknown & { [key: string]: () => Promise<unknown> }) = (
+          target[key] as ExecutionFunction
+        )(...functionArgs);
         const callbackEnding = options.callbackEnding;
         if (
           callbackEnding !== undefined &&
-          typeof syncResult[callbackEnding] === "function" &&
-          fieldValueType === "function"
+          fieldValueType === "function" &&
+          syncResult instanceof Object &&
+          typeof syncResult[callbackEnding] === "function"
         ) {
           return {
             ...syncResult,
@@ -99,15 +105,14 @@ export const proxyHandlerGenericExecution: ProxyHandlerGenericExecution = (optio
                 });
                 return asyncResult;
               } catch (error) {
-                error =
-                  options.onError({
-                    functionArgs,
-                    processingResult: ProcessingResult.failed,
-                    functionError: error,
-                    fieldValueType,
-                    fieldKey: String(key),
-                    processingStrategy: ProcessingType.callbackEnding,
-                  }) || error;
+                error = (options.onError({
+                  functionArgs,
+                  processingResult: ProcessingResult.failed,
+                  functionError: error,
+                  fieldValueType,
+                  fieldKey: String(key),
+                  processingStrategy: ProcessingType.callbackEnding,
+                }) || error) as OnErrorPayload["functionError"];
                 throw error;
               }
             },
@@ -127,15 +132,14 @@ export const proxyHandlerGenericExecution: ProxyHandlerGenericExecution = (optio
               });
               return asyncResult;
             } catch (error) {
-              error =
-                options.onError({
-                  functionArgs,
-                  processingResult: ProcessingResult.failed,
-                  functionError: error,
-                  fieldValueType,
-                  fieldKey: String(key),
-                  processingStrategy: ProcessingType.promise,
-                }) || error;
+              error = (options.onError({
+                functionArgs,
+                processingResult: ProcessingResult.failed,
+                functionError: error,
+                fieldValueType,
+                fieldKey: String(key),
+                processingStrategy: ProcessingType.promise,
+              }) || error) as OnErrorPayload["functionError"];
               throw error;
             }
           })();
@@ -150,15 +154,14 @@ export const proxyHandlerGenericExecution: ProxyHandlerGenericExecution = (optio
         });
         return syncResult;
       } catch (error) {
-        error =
-          options.onError({
-            functionArgs,
-            processingResult: ProcessingResult.failed,
-            functionError: error,
-            fieldValueType,
-            fieldKey: String(key),
-            processingStrategy: ProcessingType.synchronous,
-          }) || error;
+        error = (options.onError({
+          functionArgs,
+          processingResult: ProcessingResult.failed,
+          functionError: error,
+          fieldValueType,
+          fieldKey: String(key),
+          processingStrategy: ProcessingType.synchronous,
+        }) || error) as OnErrorPayload["functionError"];
         throw error;
       }
     };